# api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: OpenEAAP API
  description: |
    OpenEAAP (Open Enterprise AI Agent Platform) REST API Specification.

    This API provides endpoints for managing AI agents, workflows, models, and knowledge bases.

    ## Authentication
    All API requests require authentication using Bearer tokens:
    ```
    Authorization: Bearer <your-api-token>
    ```

    ## Rate Limiting
    - Default: 100 requests per minute per user
    - Burst: 200 requests per minute

    ## Versioning
    API version is included in the path: `/api/v1/...`
  version: 1.0.0
  contact:
    name: OpenEAAP Team
    email: support@openeeap.org
    url: https://github.com/openeeap/openeeap
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: https://api.openeeap.org/v1
    description: Production server
  - url: https://staging-api.openeeap.org/v1
    description: Staging server
  - url: http://localhost:8080/v1
    description: Local development server
...
paths:
...
  /workflows:
    get:
      tags: [workflows]
      summary: List all workflows
      operationId: listWorkflows
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Workflow'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
      security:
        - BearerAuth: []

    post:
      tags: [workflows]
      summary: Create a new workflow
      operationId: createWorkflow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkflowRequest'
      responses:
        '201':
          description: Workflow created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workflow'
        '400':
          $ref: '#/components/responses/BadRequest'
      security:
        - BearerAuth: []
  /workflows/{workflowId}:
    get:
      tags: [workflows]
      summary: Get workflow by ID
      operationId: getWorkflow
      parameters:
        - $ref: '#/components/parameters/WorkflowIdParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workflow'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        - BearerAuth: []

    delete:
      tags: [workflows]
      summary: Delete workflow
      operationId: deleteWorkflow
      parameters:
        - $ref: '#/components/parameters/WorkflowIdParam'
      responses:
        '204':
          description: Workflow deleted successfully
      security:
        - BearerAuth: []
  /workflows/{workflowId}/run:
    post:
      tags: [workflows]
      summary: Run workflow
      operationId: runWorkflow
      parameters:
        - $ref: '#/components/parameters/WorkflowIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RunWorkflowRequest'
      responses:
        '200':
          description: Workflow execution started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowExecution'
      security:
        - BearerAuth: []
  /workflows/{workflowId}/executions/{executionId}:
    get:
      tags: [workflows]
      summary: Get workflow execution status
      operationId: getWorkflowExecution
      parameters:
        - $ref: '#/components/parameters/WorkflowIdParam'
        - $ref: '#/components/parameters/ExecutionIdParam'
      responses:
        '200':
          description: Execution status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowExecution'
      security:
        - BearerAuth: []
  /workflows/{workflowId}/executions/{executionId}/pause:
    post:
      tags: [workflows]
      summary: Pause workflow execution
      operationId: pauseWorkflowExecution
      parameters:
        - $ref: '#/components/parameters/WorkflowIdParam'
        - $ref: '#/components/parameters/ExecutionIdParam'
      responses:
        '200':
          description: Execution paused
      security:
        - BearerAuth: []
  /workflows/{workflowId}/executions/{executionId}/resume:
    post:
      tags: [workflows]
      summary: Resume workflow execution
      operationId: resumeWorkflowExecution
      parameters:
        - $ref: '#/components/parameters/WorkflowIdParam'
        - $ref: '#/components/parameters/ExecutionIdParam'
      responses:
        '200':
          description: Execution resumed
      security:
        - BearerAuth: []

  # ==================== Model APIs ====================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeParam:
      name: page_size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    WorkflowIdParam:
      name: workflowId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ExecutionIdParam:
      name: executionId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    Workflow:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        steps:
          type: array
          items:
            $ref: '#/components/schemas/WorkflowStep'
        trigger:
          type: object
          additionalProperties: true
        status:
          type: string
          enum: [active, inactive]
        created_at:
          type: string
          format: date-time
    WorkflowStep:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        agent_id:
          type: string
          format: uuid
        condition:
          type: string
        next_steps:
          type: array
          items:
            type: string
    CreateWorkflowRequest:
      type: object
      required:
        - name
        - steps
      properties:
        name:
          type: string
        description:
          type: string
        steps:
          type: array
          items:
            $ref: '#/components/schemas/WorkflowStep'
        trigger:
          type: object
    RunWorkflowRequest:
      type: object
      properties:
        input:
          type: object
          additionalProperties: true
        context:
          type: object
          additionalProperties: true
    WorkflowExecution:
      type: object
      properties:
        id:
          type: string
          format: uuid
        workflow_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [running, paused, completed, failed]
        current_step:
          type: string
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time

    # ==================== Model Schemas ====================
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        page_size:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...